+++ packages/orq-rc/RELEASES.md
@@ -2339,3 +2339,13 @@ Based on:
 - [python v3.11.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v3.11.0-rc.16] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.16 - packages/orq-rc
+
+## 2025-08-15 12:54:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.605.0 (2.684.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.11.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v3.11.0-rc.17] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.17 - packages/orq-rc